multiboot2: don't fail on 0x9 tag when 0x3 is available
A kernel can define multiple entries. Even if they are marked as required,
it makes no sense to panic when there is a valid and expected alternative.

Further, this aligns the behavior with GRUB2.

This solution is however simple, as it expects that all tags appear in
natural order. However, every developer that is sane will order them
naturally.