nix: build Limine in Nix
This packages Limine in Nix and exports it via the Nix Flake.
Due to the nature of the project, i.e., the self-hacked
"git submodule checkout with additional patches" in `./bootstrap`,
packing this in a Nix-way is very complicated.

The big advantage of packaging Limine in Nix directly as part
of the flake (and not in nixpkgs) is that instead of using a
fixed version from nixpkgs coming from a release, one can use
the current source and build everything in Nix.